create property

+++ rooms/src/main/java/com/thuc/rooms/repository/CityRepository.java
@@ -8,9 +8,11 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface CityRepository extends JpaRepository<City, Long> {
+public interface CityRepository extends JpaRepository<City, Integer> {
 
     Optional<City> findBySlug(String slug);
 
     City findByNameContainingIgnoreCase(String destination);
+
+
 }
